% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_inflation_rates.R
\name{available_inflation_curves}
\alias{available_inflation_curves}
\title{Available inflation swap curves}
\usage{
available_inflation_curves()
}
\value{
data frame of inflation swap curves that are currently available
}
\description{
Returns a data table of current inflation swap curves that can be called in functions such as
\code{swap_implied_inflation_data}
}
\examples{
available_inflation_curves()

}
