% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_interest_rates.R
\name{available_interest_rate_curves}
\alias{available_interest_rate_curves}
\title{Available interest rate swap curves}
\usage{
available_interest_rate_curves()
}
\value{
data frame of swap curves that are currently available
}
\description{
Returns a data table of current interest rate swap curves that can be called in functions such as
\code{swap_implied_interest_rate_data}
}
\examples{
available_interest_rate_curves()

}
