% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakevens.R
\name{breakeven_level}
\alias{breakeven_level}
\title{Breakeven levels}
\usage{
breakeven_level(inflation_ISINs, nominal_ISINs, ref_date = NULL)
}
\arguments{
\item{inflation_ISINs}{character, vector of inflation linked bonds}

\item{nominal_ISINs}{character, vector of nominal bonds, same length as \code{inflation_ISINs}}

\item{ref_date}{date, a reference date for the yield level, defaults to \code{Sys.Date()}}
}
\value{
data frame of inflation ISINs paired with nominal ISINs and the breakeven level
}
\description{
Returns inflation-linked bond breakeven level
}
\details{
takes a list of inflation linked bonds and their nominal comparitor bonds
ISINs to calculate the breakeven level
}
\examples{
   \dontrun{
       breakeven_level(
         inflation_ISINs = c("US912810PZ57", "US912810QP66"),
         nominal_ISINs = c("US91282CDP32", "US912810QN19"),
         ref_date = Sys.Date()
       )
   }
}
