% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breakevens.R
\name{breakeven_pairing}
\alias{breakeven_pairing}
\title{Breakeven pairings}
\usage{
breakeven_pairing(
  inflation_ISINs,
  nominal_ISINs,
  tie_method = "issue_date",
  ref_date = NULL
)
}
\arguments{
\item{inflation_ISINs}{character, vector of inflation linked bonds to find the breakeven comparitor for}

\item{nominal_ISINs}{character, vector of nominal bonds to pair the inflation linked bonds against}

\item{tie_method}{character, currently has two options to break ties when more than one bond has the closest maturity.
\code{"issue_date"} chooses the nominal bond with the closest similar maturity. \code{"high_yield"} chooses the nominal
bond with the highest yield, effectively the cheapest.}

\item{ref_date}{date, a reference date for the outstanding bonds, yield level etc, defaults to \code{Sys.Date()}}
}
\value{
data frame of inflation ISINs paired with nominal ISINs
}
\description{
Returns inflation-linked bond paired with closest nominal comparitor
}
\details{
returns a tibble of inflation linked ISINs paired with a nominal to imply breakevens.
Works with
}
\examples{
   \dontrun{
       breakeven_pairing(
         inflation_ISINs = "US912810QP66",
         nominal_ISINs = c("US91282CDP32", "US912810QN19", "US912810QE10"),
         ref_date = Sys.Date()
       )
   }
   
}
