% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_func.R
\name{calc_beta_ts}
\alias{calc_beta_ts}
\title{Calculate rolling beta of two time series}
\usage{
calc_beta_ts(
  dependent,
  independent,
  window = 180,
  method = "change",
  change_window = 5,
  include_r2 = FALSE
)
}
\arguments{
\item{dependent}{Dependent variable. Object of class data.frame, with column names 'Date' and 'Value'}

\item{independent}{Independent variable. Object of class data.frame, with column names 'Date' and 'Value'}

\item{window}{window to calculate beta on a rolling basis. The time horizon of rolling window for the beta in days}

\item{method}{method to calculate beta, choice are "change", "level", "percent_change", "log_change", "log_level"}

\item{change_window}{is the time horizon for the change measurement, eg 5 would be a weekly change, 22 a monthly}

\item{include_r2}{include R squared in the output series}
}
\value{
long formatted data frame of beta values calculated on a rolling basis
}
\description{
Allows you to calculate rolling beta realtionship between independent and dependent variables over given window
}
\details{
function expected two (time series) data frames, in long format from which the rolling beta/relationship is calculated
}
\examples{
\dontrun{ 
    calc_beta_ts(independent, dependent
                 window = 90,
                 method = "change",
                 change_window = 5,
                 include_r2 = TRUE)
  }

}
