% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_func.R
\name{calc_beta_ts_bloomberg}
\alias{calc_beta_ts_bloomberg}
\title{Calculate rolling beta of two time series}
\usage{
calc_beta_ts_bloomberg(
  dependent_security,
  dependent_field,
  independent_security,
  independent_field,
  to_date = Sys.Date(),
  from_date = Sys.Date() - 365,
  window = 180,
  method = "level",
  change_window = 5,
  include_r2 = FALSE
)
}
\arguments{
\item{dependent_security}{Dependent security ticker, object of class character, e.g. "USGGBE03 Index"}

\item{dependent_field}{Dependent security bloomberg field, object of class character, e.g. "PX_LAST"}

\item{independent_security}{Independent security ticker, object of class character, e.g. "USGGBE03 Index"}

\item{independent_field}{Independent security bloomberg field, object of class character, e.g. "PX_LAST"}

\item{to_date}{date object, end date of the series calculation}

\item{from_date}{date object, start date of the series calculation}

\item{window}{window to calculate beta on a rolling basis. The time horizon of rolling window for the beta in days}

\item{method}{method to calculate beta, choice are "change", "level", "percent_change", "log_change", "log_level"}

\item{change_window}{is the time horizon for the change measurement, eg 5 would be a weekly change, 22 a monthly}

\item{include_r2}{include R squared in the output series}
}
\value{
long formatted data frame of beta values calculated on a rolling basis
}
\description{
Allows you to calculate rolling beta realtionship between independent and dependent variables over given window
}
\details{
function expected two (time series) data frames, in long format from which the rolling beta/relationship is calculated
}
\examples{
\dontrun{
calc_beta_ts_bloomberg(dependent_security = "USGGBE03 Index", 
dependent_field = "PX_LAST", 
independent_security = "CL1 Comdty", 
independent_field = "PX_LAST",
window = 90,
method = "change",
change_window = 5,
include_r2 = TRUE)
}

}
