% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{cashflow_table}
\alias{cashflow_table}
\title{Cashflow table generator}
\usage{
cashflow_table(
  maturity_date,
  coupon,
  payment_frequency,
  ref_date = Sys.Date(),
  settle_convention = 1,
  day_count = "ACT/ACT"
)
}
\arguments{
\item{maturity_date}{date, the date of the maturity ie final cashflow}

\item{coupon}{numeric, the coupon in percent to be paid at regular intervals}

\item{payment_frequency}{numeric, the number of times per year the coupon is to be paid}

\item{ref_date}{date, the expected trade date, default to today}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{day_count}{character, the day count convention of the instrument. Can currently only be \code{"ACT/ACT"}.}
}
\value{
a tibble of cashflows with discount periods
}
\description{
Gernerates a table of cashflows for a fixed income instrument with regular cashflows, including
the discounting fraction of each period
}
\examples{
cashflow_table(as.Date("2023-02-21"), 0.1, 2)


}
