% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portfolio_functions.R
\name{cashflow_table_ilb_portfolio}
\alias{cashflow_table_ilb_portfolio}
\title{ILB portfolio cashflows}
\usage{
cashflow_table_ilb_portfolio(
  bond_ISINs,
  ref_date = Sys.Date(),
  inflation_adj = TRUE,
  seasonally_adjust = FALSE,
  sa_method = "linear",
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{bond_ISINs}{character, string or vector of strings of inflation linked bond ISINs e.g. US912810SV17}

\item{ref_date}{date, calculation date for cashflows, only future cashflows shown}

\item{inflation_adj}{logical, whether the cashflow should include the inflaton adjusted cashflows}

\item{seasonally_adjust}{logical, seasonal adjustment}

\item{sa_method}{method for seasonal adjustment, defaults to 'linear'}

\item{lm_database_path}{Path class character to list management database}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
data frame of cashflows for the bonds
}
\description{
Returns a data table of future cashflows for a portfolio of ISINs. Can be either inflation adjusted
or unadjusted. Assumes 1mn notional (face value), so that adjustments for holding size will have to be done
after calling the function.

Built on the back of \link[InflationTools]{cashflow_table} and  \link[InflationTools]{linker_cashflow_table}
functions.
}
\examples{
\dontrun{  
  cashflow_table_ilb_portfolio(bond_ISINs = c("US91282CCM10", "US91282CDC29"))
}

}
