% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{convexity_inflation}
\alias{convexity_inflation}
\title{Calculate convexity for an inflation linked instrument}
\usage{
convexity_inflation(linker_cashflow_table, real_yield, payment_frequency)
}
\arguments{
\item{linker_cashflow_table}{data.table, table of inflation linked (adjusted) cashflows and discount periods, can be generated by
\link[InflationTools]{linker_cashflow_table}.}

\item{real_yield}{numeric, real yield of instrument at reference date}

\item{payment_frequency}{numeric, payment frequency per annum of the instrument eg 2 is semi-annual}
}
\value{
numeric, the convexity
}
\description{
Using the cashflows from an inflation linked instrument, calculate the the convexity (also known as the second derivative
of the interest rate risk).
}
\examples{
temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
linker_cash_tbl <- linker_cashflow_table(temp_cashflow_tbl,
                                         exampleUSInflationCurve,
                                         260)
convexity_inflation(linker_cash_tbl,
                               -1,
                               2)
                               
                                   
}
