% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{exampleExposurePortfolio}
\alias{exampleExposurePortfolio}
\title{Fake portfolio of trades}
\format{
A data frame of 5 rows and 6 columns
\describe{
\item{instrument_type}{character,  whether the instrument is a inflation linked bond, zero coupon swap or asset swap.}
\item{maturity_date}{date, the maturity date of the instrument}
\item{notional}{numeric, the notional of the instrument}
\item{payment_frequency}{numeric, payment frequency of the instrument, eg 2 is semi-annual}
\item{coupon}{numeric, coupon (annualised) of the instrument}
\item{index_ratio}{numeric, the index ratio of the instrument on the day of calculation}
}
}
\source{
SH
}
\usage{
exampleExposurePortfolio
}
\description{
A datatable of fake trades and trade dates for use with the exposure functions.
}
\keyword{datasets}
