% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{exampleUSInflationCurve}
\alias{exampleUSInflationCurve}
\title{Projected US inflation curve (July 2021)}
\format{
A data frame of 387 rows and 3 columns
\describe{
\item{date}{date, the date of the index value}
\item{proj_infl}{numeric, the projected (or realised for historic) inflation index value on the date}
\item{season_factor_cuml}{numeric, the multiplicative seasonal adjustment applied to the index value}
}
}
\source{
Bloomberg and package calculations
}
\usage{
exampleUSInflationCurve
}
\description{
A datatable with the realised and projected US CPURNSA inflation curve from Jan 2019 until Mar 2051
}
\keyword{datasets}
