% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_db_functions.R
\name{get_inflation_series_db}
\alias{get_inflation_series_db}
\title{Query inflation index series from inflation database}
\usage{
get_inflation_series_db(
  db_series,
  from = "1990-01-01",
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{db_series}{Inflation index series name, e.g. CPTFEMU Index}

\item{from}{Object of class date, the start date for the series}

\item{infl_database_path}{Path to inflation database, default is the rminfldbpath}
}
\value{
inflation index series
}
\description{
Query and return inflation index series from inflation database.
}
\details{
get_inflation_series_db takes an inflation index ticker and returns data series
}
\examples{

get_inflation_series_db("CPTFEMU Index")

}
