% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_db_functions.R
\name{historic_inflation_curve_db}
\alias{historic_inflation_curve_db}
\title{Historic inflation curve from database}
\usage{
historic_inflation_curve_db(
  inflation_index,
  ref_date = NULL,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{inflation_index}{character, the name of the inflation index, with bloomberg suffix, eg \code{"CPURNSA Index"}}

\item{ref_date}{date, the date of the historic curve}

\item{infl_database_path}{charracter, path to inflation database}
}
\value{
data.table, historic projected inflation curve and associated meta data
}
\description{
Pull historic projected inflation curves from the database, does not require a bloomberg connection
}
\examples{
historic_inflation_curve_db("CPTFEMU Index", 
                            ref_date = NULL,
                            infl_database_path = rminfldbpath)

}
