% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_inflation_rates.R
\name{implied_proj_inflation_curve}
\alias{implied_proj_inflation_curve}
\title{Implied Projected Inflation Curve}
\usage{
implied_proj_inflation_curve(
  implied_curve,
  base_level,
  base_month = lubridate::month(Sys.Date()),
  base_year = lubridate::year(Sys.Date()),
  frequency = 12
)
}
\arguments{
\item{implied_curve}{Object of class data frame, specifying the market implied curve, expects column names 'implied_infl', 'maturity' both numeric}

\item{base_level}{Base level of the inflation index, numeric}

\item{base_month}{Numeric object, base month for the inflation index}

\item{base_year}{Numeric object, base year for the inflation index}

\item{frequency}{frequency of the index, e.g. 12 for monthly}
}
\value{
data frame of projections implied from the curve
}
\description{
Returns inflation projected from implied inflation curve
}
\details{
Implied curve can be taken from zero-coupon swap curve or elsewhere, from which the (marked) implied inflation can be backed out.
See also \link[InflationTools]{implied_proj_inflation_curve_db}.
}
\examples{
  implied_proj_inflation_curve(readr::read_csv("maturity,implied_infl
                                                1,2
                                                2,3
                                                3,2.5
                                                5,2"), 
                               100, 2, 2020, 12)


}
