% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_inflation_rates.R
\name{implied_proj_inflation_curve_db}
\alias{implied_proj_inflation_curve_db}
\title{Implied Projected Inflation Curve from inflation database}
\usage{
implied_proj_inflation_curve_db(
  inflation_index,
  seasonally_adjust = FALSE,
  sa_method = "linear",
  from_date = as.Date("2019-01-01"),
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{inflation_index}{Object of class character, specifying the inflation index to use}

\item{seasonally_adjust}{boolean argument, seasonal adjustment}

\item{sa_method}{method for seasonal adjustment, defaults to 'linear'}

\item{from_date}{Object of class date, the start date for the series}

\item{infl_database_path}{character, path of inflation database, default is rminfldbpath from package}
}
\value{
data frame of projections implied from the curve
}
\description{
Returns inflation projected from implied inflation curve
}
\details{
Implied curve can be taken from zero-coupon swap curve or elsewhere, from which the (marked) implied inflation can be backed out
See also \link[InflationTools]{implied_proj_inflation_curve}.
}
\examples{
\dontrun{
  implied_proj_inflation_curve_db("CPURNSA Index")
}

}
