% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_supply.R
\name{infl_link_supply_ts_db}
\alias{infl_link_supply_ts_db}
\title{Linker Supply from Database}
\usage{
infl_link_supply_ts_db(
  country,
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{country}{Object of class character, specifying country, e.g. United States = "US"}

\item{lm_database_path}{Path to list management database}

\item{infl_database_path}{Path to inflation database, default is rminfldbpath}
}
\value{
data frame of inflation-linked bond supply for a given country
}
\description{
Returns inflation-linked bond supply from inflation database for a given country
}
\details{
returns time series of inflation-linked bond supply for a given country, including notional and PV01 statistics
}
\examples{
\dontrun{
  infl_link_supply_ts_db(country = "US")
}
}
