% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_supply.R
\name{infl_linked_fed_purchases}
\alias{infl_linked_fed_purchases}
\title{Get TIPS Fed SOMA holdings data}
\usage{
infl_linked_fed_purchases(
  path = "https://markets.newyorkfed.org/api/soma/tsy/get/monthly.csv",
  start_date = Sys.Date() - 10 * 365
)
}
\arguments{
\item{path}{Object of class character, the location of the data}

\item{start_date}{start date for the series, e.g "2018-01-01"}
}
\value{
data frame of Fed holdings
}
\description{
Retrieves data frame of Fed SOMA holdings of TIPS securities, from as early as 2003
}
\details{
returns data frame of Fed holdings
}
\examples{
\dontrun{
  InflationTools::infl_link_fed_purchases(start_date = "2010-01-01")
}

}
