% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_risk_premia.R
\name{inflation_risk_premia}
\alias{inflation_risk_premia}
\title{Inflation risk premia modelled estimation}
\usage{
inflation_risk_premia(
  maturity_input,
  start_date = Sys.Date() - 8 * 365,
  regression_window = 200,
  rolling = TRUE
)
}
\arguments{
\item{maturity_input}{Object of class numeric, the maturity of the breakeven inflation (e.g. 5, 10, 30)}

\item{start_date}{Object of class date, the start date for the estimation}

\item{regression_window}{numeric, rolling window length for the regression}

\item{rolling}{Boolean, if TRUE (default), runs a rolling regression. If FALSE, runs a simple linear regression.}
}
\value{
data frame of projections of breakeven inflation, inflation risk premia, liquidity factors and residuals.
}
\description{
Returns modelled values for inflation risk premia, for a given maturity date
}
\details{
Risk premia is estimate using a set of liquidity factors (VIX, MOVE, bid-offer) and inflation expectations, the inflation risk premia
is the breakeven inflation not explained by liquidity or expectations
}
\examples{
\dontrun{
  inflation_risk_premia(maturity_input = 10, 
                        start_date = Sys.Date() - 8*365, 
                        regression_window = 200)
}

}
