% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{package}
\name{InflationTools}
\alias{InflationTools}
\alias{InflationTools-package}
\title{InflationTools package}
\description{
Fixed income inflation tools and analysis
}
\author{
\strong{Maintainer}: Sam H \email{325061@bankofengland.co.uk}

Authors:
\itemize{
  \item Josh A \email{328576@bankofengland.co.uk} [contributor]
}

}
