% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_curve_functions.R
\name{interpolate_inflation_curve}
\alias{interpolate_inflation_curve}
\title{Find the index value on a given day}
\usage{
interpolate_inflation_curve(
  projected_inflation_curve,
  ref_date = Sys.Date(),
  inflation_lag = 3,
  settle_convention = 1,
  round_5 = TRUE,
  interpolation = "daily"
)
}
\arguments{
\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl},
eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{ref_date}{date, the expected trade date, default to today}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{round_5}{logical, whether the index value should be rounded to five decimal places}
}
\value{
numeric, the interpolated index value for ref_date
}
\description{
Find the index value on a given day
}
\examples{

interpolate_inflation_curve(exampleUSInflationCurve, 
                            ref_date = as.Date("2021-05-01"),
                            inflation_lag = 3)


}
