% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{interpolate_projected_inflation}
\alias{interpolate_projected_inflation}
\title{Interpolate inflation for instrument cashflow}
\usage{
interpolate_projected_inflation(
  cashflow_table,
  projected_inflation_curve,
  ref_date = Sys.Date(),
  settle_convention = 1,
  inflation_lag = 3,
  interpolation = "daily"
)
}
\arguments{
\item{cashflow_table}{data.table, table of cashflows and discount periods, can be generated by \link[InflationTools]{cashflow_table}}

\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl},
eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{ref_date}{date, the expected trade date, default to today}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{interpolation}{character, can be daily or monthly, inflation interpolation method}
}
\value{
a cashflow tibble with interpolated inflation indexes
}
\description{
For an instrument with regular cashflows, interpolate a projected (or realised) inflation curve to get the index
values for the cashflow dates.
}
\examples{

temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
interpolate_projected_inflation(temp_cashflow_tbl,
                                exampleUSInflationCurve)


}
