% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{interpolation_reference_date}
\alias{interpolation_reference_date}
\title{Lagged reference date for interpolation}
\usage{
interpolation_reference_date(
  ref_date,
  settle_convention = 1,
  inflation_lag = 3,
  interpolation = "daily"
)
}
\arguments{
\item{ref_date}{date, the date to be lagged}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{interpolation}{character, can be \code{daily} or \code{monthly}, inflation interpolation method}
}
\value{
date
}
\description{
For an instrument using standard inflation lag, gives the interpolation date based on
the input date, lag and interpolation method.
}
\examples{

interpolation_reference_date(as.Date("2027-02-21"))
interpolation_reference_date(as.Date("2032-07-15"),
                             interpolation = "monthly")



}
