% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOTA_spreads.R
\name{iota_asw_fixed_maturity}
\alias{iota_asw_fixed_maturity}
\title{ASW iota spread at a constant maturity point}
\usage{
iota_asw_fixed_maturity(
  maturity,
  inflation_ISINs,
  nominal_ISINs,
  spread = 0.5,
  ref_date = NULL,
  tie_method = "issue_date"
)
}
\arguments{
\item{maturity}{integer, constant maturity point in years}

\item{inflation_ISINs}{character, vector of inflation linked bonds}

\item{nominal_ISINs}{character, vector of nominal bonds to pair the inflation linked bonds against}

\item{spread}{numeric, difference in years either side of the maturity point to chose bonds for
regression, eg if maturity is 5 and spread 0.5 then the range will be 4.5y to 5.5y}

\item{ref_date}{date, a reference date for the yield level, defaults to \code{Sys.Date()}}

\item{tie_method}{character, currently has two options to break ties when more than one bond has the closest maturity.
\code{"issue_date"} chooses the nominal bond with the closest similar maturity. \code{"high_yield"} chooses the nominal
bond with the highest yield, effectively the cheapest. From \code{breakeven_pairing}.}
}
\value{
numeric, ASW iota spread level
}
\description{
Returns iota spread for a maturity point on the iota curve on a set day
}
\details{
Linear regression on iota spreads of bonds of a spread around the chosen maturity, in order
to get a constant maturity style iota spread
}
\examples{
   \dontrun{
       iota_asw_fixed_maturity(
         maturity = 5,
         inflation_ISINs = c("US912810PZ57", "US912810QP66"),
         nominal_ISINs = c("US91282CDP32", "US912810QN19"),
         ref_date = Sys.Date()
       )
   }
}
