% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOTA_spreads.R
\name{iota_asw_level}
\alias{iota_asw_level}
\title{Iota ASW level}
\usage{
iota_asw_level(inflation_ISINs, nominal_ISINs, ref_date = NULL)
}
\arguments{
\item{inflation_ISINs}{character, vector of inflation linked bonds to find the breakeven comparitor for}

\item{nominal_ISINs}{character, vector of nominal bonds to pair the inflation linked bonds against}

\item{ref_date}{date, a reference date for the yield level, defaults to \code{Sys.Date()}}
}
\value{
data frame of inflation ISINs paired with nominal ISINs and the breakeven level
}
\description{
Returns inflation-linked bond par-par ASW iota levels
}
\details{
takes a list of inflation linked bonds and their nominal comparitor bonds
ISINs to calculate the iota spread using
}
\examples{
   \dontrun{
       iota_asw_level(
         inflation_ISINs = c("US912810PZ57", "US912810QP66"),
         nominal_ISINs = c("US91282CDP32", "US912810QN19"),
         ref_date = Sys.Date()
       )
   }
}
