% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOTA_spreads.R
\name{iota_asw_level_db}
\alias{iota_asw_level_db}
\title{Iota ASW levels from lm database}
\usage{
iota_asw_level_db(
  inflation_list,
  nominal_list,
  ref_date = NULL,
  tie_method = "issue_date",
  lm_database_path = rmlmdbpath
)
}
\arguments{
\item{inflation_list}{character, list in list management database for inflation bonds}

\item{nominal_list}{character, list in list management database for nominal bonds}

\item{ref_date}{date, a reference date for the yield level, defaults to \code{Sys.Date()}}

\item{tie_method}{character, currently has two options to break ties when more than one bond has the closest maturity.
\code{"issue_date"} chooses the nominal bond with the closest similar maturity. \code{"high_yield"} chooses the nominal
bond with the highest yield, effectively the cheapest. From \code{breakeven_pairing}.}

\item{lm_database_path}{Path class character to list management database}
}
\value{
data frame of inflation ISINs paired with nominal ISINs and the breakeven level
}
\description{
Returns inflation-linked bond breakeven level
}
\details{
takes a list of inflation linked bonds and their nominal comparitor bonds
}
\examples{
   \dontrun{
       iota_asw_level_db(
         inflation_list = "USD_Infl_US",
         nominal_list = "USD_Govt_US",
         ref_date = Sys.Date()
       )
   }


}
