% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOTA_spreads.R
\name{iota_spreads_bloomberg}
\alias{iota_spreads_bloomberg}
\title{Get IOTA spreads (US)}
\usage{
iota_spreads_bloomberg(start_date = Sys.Date() - 365, end_date = Sys.Date())
}
\arguments{
\item{start_date}{start date for the series, e.g "2018-01-01"}

\item{end_date}{end date for the series, e.g "2018-01-01"}
}
\value{
data frame of IOTA breakeven spreads
}
\description{
Retrieves IOTA breakeven spreads from Bloomberg
}
\details{
returns data frame of IOTA breakevens, across maturities
}
\examples{
\dontrun{
  InflationTools::iota_spreads_bloomberg(start_date = "2010-01-01",  end_date = Sys.Date())
}

}
