% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_curve_functions.R
\name{lag_inflation_curve}
\alias{lag_inflation_curve}
\title{Lag the inflation curve by a number of months}
\usage{
lag_inflation_curve(inflation_curve, inflation_lag = 3)
}
\arguments{
\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{projected_inflation_curve}{data.frame, curve of projected inflation with columns \code{date} and \code{proj_infl},
eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}
}
\value{
data.frame, the input data.frame with a changed date column to reflect the relevant lag
}
\description{
Lags an inflation curve by a set number of months. In inflation calculations the end month inflation value is
used as the value for the first day of the month. So a three month lag is more like a two month and one day lag.
}
\examples{

lag_inflation_curve(exampleUSInflationCurve, 
                            inflation_lag = 3)


}
