% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{linker_cashflow_table}
\alias{linker_cashflow_table}
\title{Inflation linked bond cashflow table generator}
\usage{
linker_cashflow_table(
  cashflow_table,
  projected_inflation_curve,
  reference_cpi,
  ref_date = Sys.Date(),
  settle_convention = 1,
  inflation_lag = 3,
  principal_floor = 100,
  coupon_floor = 0
)
}
\arguments{
\item{cashflow_table}{data.table, table of cashflows and discount periods, can be generated by \code{cashflow_table}}

\item{projected_inflation_curve}{data.table, curve of projected inflation can be generated by \code{implied_proj_inflation_curve}}

\item{reference_cpi}{numeric, the reference value of the inflation linked bond at issue}

\item{ref_date}{date, the expected trade date, default to today}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{principal_floor}{numeric, the percentage floor of the principal, eg 100 is floored at par}

\item{coupon_floor}{numeric, the percentage floor of the coupon, eg 0 is not floored}
}
\value{
a tibble of both inflation adjusted and unadjusted cashflows with discount periods
}
\description{
Adjusts the cashflows from \code{cashflow_table} for an inputted projected inflation curve. Also accounts
for floors on the principal and coupon.
}
\examples{
temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
linker_cashflow_table(temp_cashflow_tbl,
                      exampleUSInflationCurve,
                      260)
                                
                                
                                
}
