% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curve_functions.R
\name{linker_curve_data_db}
\alias{linker_curve_data_db}
\title{Linker Curve from Database}
\usage{
linker_curve_data_db(
  country,
  projected_inflation_curve = NULL,
  ref_date = Sys.Date(),
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{country}{Object of class character, specifying country, e.g. United States = "US"}

\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl}
, eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{ref_date}{date, trade or calculation date, regular settle is assumed}

\item{lm_database_path}{Path to list management database}

\item{infl_database_path}{Path to inflation database}
}
\value{
data frame of yield measures
}
\description{
Returns inflation-linked bonds from inflation database for a given country
}
\details{
returns universe of linkers on the real yield curve for a given country, and related levels including seasonally adjusted real yields
}
\examples{
\dontrun{
   linker_curve_data_db(country = "US", 
                        ref_date = Sys.Date())
}

}
