% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILB_info_calculations.R
\name{linker_info_func}
\alias{linker_info_func}
\alias{linker_info_yield}
\alias{linker_info}
\alias{linker_info_db}
\alias{linker_info_price}
\title{Linker statistics calculations}
\usage{
linker_info_yield(
  real_yield,
  projected_inflation_curve,
  reference_cpi,
  inflation_lag,
  principal_floor,
  coupon_floor,
  maturity_date,
  coupon,
  payment_frequency,
  ref_date = Sys.Date(),
  settle_convention = 1,
  day_count = "ACT/ACT"
)

linker_info(
  real_yield = NULL,
  real_price = NULL,
  projected_inflation_curve,
  reference_cpi,
  inflation_lag,
  principal_floor,
  coupon_floor,
  maturity_date,
  coupon,
  payment_frequency,
  ref_date = Sys.Date(),
  settle_convention = 1,
  day_count = "ACT/ACT"
)

linker_info_db(
  bond_ISIN,
  projected_inflation_curve,
  real_price = NULL,
  real_yield = NULL,
  ref_date = Sys.Date(),
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)

linker_info_price(
  real_price,
  projected_inflation_curve,
  reference_cpi,
  inflation_lag,
  principal_floor,
  coupon_floor,
  maturity_date,
  coupon,
  payment_frequency,
  ref_date = Sys.Date(),
  settle_convention = 1,
  day_count = "ACT/ACT"
)
}
\arguments{
\item{real_yield}{numeric, real yield of bond at reference date in percent}

\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl}
, eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{reference_cpi}{numeric, the CPI index of the bond at issue, sometimes also called "base CPI"}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{principal_floor}{numeric, the percentage floor of the principal, eg 100 is floored at par}

\item{coupon_floor}{numeric, the percentage floor of the coupon, eg 0 is not floored}

\item{maturity_date}{date, the date of the maturity ie final cashflow}

\item{coupon}{numeric, the coupon in percent to be paid at regular intervals}

\item{payment_frequency}{numeric, the number of times per year the coupon is to be paid}

\item{ref_date}{date, trade or calculation date, regular settle is assumed in \code{_db} functions}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{day_count}{character, the day count convention of the instrument. Can currently only be \code{"ACT/ACT"}.}

\item{real_price}{numeric, clean price of bond at reference date}

\item{bond_ISIN}{character, specifying bond ISIN of an inflation linked bond, e.g. \code{"US912810SV17"}, must be in the database}

\item{lm_database_path}{Path class character to list management database}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
list of bond statistics
}
\description{
Returns information on an inflation-linked bond given either a price or yield, for 1mn local currency nominal. Using either
an ISIN with the database (\code{_db}) or with the full set of information about the bond.
One of either a real yield or price is required in order to work out the risk values.
}
\examples{
\dontrun{
   linker_info_db(bond_ISIN = "US912810SV17", 
               projected_inflation_curve = exampleUSInfaltionCurve, 
               real_yield = -0.3)
   linker_info_db(bond_ISIN = "US912810SV17", 
               projected_inflation_curve = exampleUSInfaltionCurve, 
               real_price = 120)
}
}
