% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{linker_lm_info}
\alias{linker_lm_info}
\title{Linker info from RM list management database}
\format{
A data frame of 135 rows and 8 columns
\describe{
\item{SecurityDes}{character, the name of the linker}
\item{ISIN}{character, the ISIN of the linker plus the bloomberg suffix eg Govt, Corp}
\item{ListName}{character, the list in the list management database the linker is in}
\item{REFERENCE_INDEX}{character, the bloomberg code for the inflation index the linker references}
\item{INFLATION_LAG}{numeric, the lag of the linker accrual to the reference index}
\item{StartDate}{date, the date the linker was added to the list management database}
\item{EndDate}{date, the date the linker will leave the list management database}
\item{MaturityDate}{date, the maturity date of the linker}
}
}
\source{
RM list_management database
}
\usage{
linker_lm_info
}
\description{
A dataset of inflation linked bonds their ISINs, maturities, reference indexes and lags.
}
\keyword{datasets}
