% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ILB_other_calculations.R
\name{linker_proceeds_yield}
\alias{linker_proceeds_yield}
\title{Calculate linker proceeds}
\usage{
linker_proceeds_yield(
  real_yield,
  projected_inflation_curve,
  reference_cpi,
  inflation_lag,
  maturity_date,
  coupon,
  payment_frequency,
  ref_date = Sys.Date(),
  settle_convention = 1,
  day_count = "ACT/ACT"
)
}
\arguments{
\item{real_yield}{numeric, real yield of bond at reference date in percent}

\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl}
, eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{reference_cpi}{numeric, the CPI index of the bond at issue, sometimes also called "base CPI"}

\item{inflation_lag}{numeric, the lag in months that instrument has to the inflation index}

\item{maturity_date}{date, the date of the maturity ie final cashflow}

\item{coupon}{numeric, the coupon in percent to be paid at regular intervals}

\item{payment_frequency}{numeric, the number of times per year the coupon is to be paid}

\item{ref_date}{date, trade or calculation date, regular settle is assumed in \code{_db} functions}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{day_count}{character, the day count convention of the instrument. Can currently only be \code{"ACT/ACT"}}
}
\value{
numeric, the proceeds amount
}
\description{
Calculate the proceeds of an inflation linked bond on a day given the yield, for notional
of 100mn local currency.
}
\examples{
 \dontrun{
     linker_proceeds_yield(real_yield = -2.5, 
                           projected_inflation_curve = exampleUSInflationCurve, 
                           reference_cpi = 255,
                           inflation_lag = 3, 
                           maturity_date = as.Date("2030-10-15"), 
                           coupon = 0.25,
                           payment_frequency = 2)
 }

}
