% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdashboard_functions.R
\name{linker_risk_dashboard}
\alias{linker_risk_dashboard}
\title{Openlink report linker risk calculation}
\usage{
linker_risk_dashboard(
  instrument_ticker,
  maturity_date,
  clean_price,
  nominal,
  currency,
  ref_date = Sys.Date(),
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{instrument_ticker}{character, the openlink ticker, normally of the format \code{"issuer_coupon_date"}}

\item{maturity_date}{date, the date of the maturity ie final cashflow}

\item{clean_price}{numeric, the real (or clean price) of an inflation linked bond}

\item{nominal}{numeric, the face value or nominal of the instrument traded}

\item{currency}{character, the currency denomination of the instrument}

\item{lm_database_path}{Path class character to list management database}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
list, includes DV01, IE01 and any error messages
}
\description{
Calculate the linker risk associated with holdings from information found in the openlink report
}
\details{
Runs using the more limited information provided in the daily openlink P&L reports. Is supplemented
by inflation and list management databases for static data, so it doesn't need to connect to bloomberg.
Probably shouldn't be used outside of openlink with \link[InflationTools]{linker_info_func} doing the
same calculations with less input required.
}
\examples{
 \dontrun{
     linker_risk_dashboard(instrument_ticker = "TII_0.5_151025", 
                           maturity_date = as.Date("2025-10-15"), 
                           clean_price = 107,
                           nominal = 120000000, 
                           currency = "USD")
 }
}
