% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_db_functions.R
\name{linker_static_all_db}
\alias{linker_static_all_db}
\title{Get all database static data}
\usage{
linker_static_all_db(
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{lm_database_path}{Path class character to list management database}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
data.table, static data on inflation linked bonds from the database
}
\description{
Get all database static data
}
\examples{
\dontrun{
  linker_static_all_db() 
}


}
