% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inflation_db_functions.R
\name{linker_static_db}
\alias{linker_static_db}
\title{Inflation linked bond static information}
\usage{
linker_static_db(
  bond_ISIN,
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{bond_ISIN}{character, string or vector of strings of inflation linked bond ISINs e.g. US912810SV17}

\item{lm_database_path}{Path to list maangement database, default is rmlmdbpath}

\item{infl_database_path}{Path to inflation database, default is the rminfldbpath}
}
\value{
dataframe of static data for inflation linked bonds
}
\description{
Query and return inflation static information for inflation linked bonds from the inflation database.
}
\details{
linker_static_db takes inflation linked bond ISINs and returns a data frame. If bond_ISIN arguement is
set to \code{NULL} will return all data.
}
\examples{
\dontrun{ 
  linker_static_db(bond_ISIN = "US912810SV17")
}

}
