% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_adjustment_cpi.R
\name{monthplot_db}
\alias{monthplot_db}
\title{Plot monthly inflation}
\usage{
monthplot_db(
  infl_index,
  end_date = Sys.Date(),
  start_date = as.Date("1980-01-01"),
  outliers_shift = c("ao2009.Oct", "ao2009.Sep"),
 
    database_path = "\\\\\\\\markets\\\\DATA\\\\DATA\\\\Offdata\\\\RM\\\\_Data\\\\Inflation\\\\inflation.db"
)
}
\arguments{
\item{infl_index}{Inflation index series name, e.g. CPTFEMU Index}

\item{end_date}{Object of class date, the end date for the series}

\item{start_date}{Object of class date, the start date for the series}

\item{outliers_shift}{Vector of characters, of any outliers (ao) or level shifts expected in the data. The \code{seas}
function from the seasonal package will identify these automatically, but they can also be pre-specified if the user
has prior knowledge of the data.}

\item{database_path}{Path to inflation database, default ////markets//DATA//DATA//Offdata//RM//_Data//Inflation//inflation.db}
}
\value{
Monthly inflation plot
}
\description{
The monthplot shows the changing trend in seasonality in different months over a number of years.
}
\details{
Plots monthly inflation for a given inflation index series
}
\examples{

monthplot_db("CPURNSA Index")


}
