% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{nominal_equivalent_yield}
\alias{nominal_equivalent_yield}
\title{Calculate nominal equivalent yield for a linker}
\usage{
nominal_equivalent_yield(
  linker_cashflow_table,
  real_price,
  payment_frequency,
  real_accrued,
  principal
)
}
\arguments{
\item{linker_cashflow_table}{data.table, table of inflation linked (adjusted) cashflows and discount periods, can be generated by
\link[InflationTools]{linker_cashflow_table}.}

\item{real_price}{numeric, the real (or clean price) of an inflation linked bond}

\item{payment_frequency}{numeric, payment frequency per annum of the instrument eg 2 is semi-annual}

\item{real_accrued}{numeric,interest accrued by an instrument since the last cashflow, can be generated by \link[InflationTools]{real_accrued}.}

\item{principal}{numeric, the principal of the /bond ie dirty proceeds less accrued}
}
\value{
numeric, a yield
}
\description{
For an inflation linked bond, calculates, based on the project inflation curve, a nominal equivalent yield for
the inflation linked instrument. It does this by taking discounting the instruments cashflow after they have been adjusted
for projected inflation, is the same way you would calculate the yield of a nominal bond. The cashflow profile of the inflation
linked instrument is skewed longer than a nominal comparitor but this shoudl not decrease the relevance of the results.
}
\examples{
temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
linker_cash_tbl <- linker_cashflow_table(temp_cashflow_tbl,
                                         exampleUSInflationCurve,
                                         260)
nominal_equivalent_yield(linker_cash_tbl,
                               145,
                               2,
                               2500, 
                               156000000)    
                               
}
