% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdashboard_functions.R
\name{ol_ticker_to_ISIN}
\alias{ol_ticker_to_ISIN}
\title{Openlink ILB ticker to ISIN}
\usage{
ol_ticker_to_ISIN(
  openlink_tickers,
  lm_database_path = rmlmdbpath,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{openlink_tickers}{come of the form (insturment)\emph{(coupon)}(maturity date) eg TII_0.25_151028. This can be unpacked and compared to
a list of outstanding inflation linked bonds to get the ISIN in question.}

\item{lm_database_path}{Path class character to list management database}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
data.frame, of ol tickers and corresponding ISINs
}
\description{
For openlink tickers inflation linked bond tickers, return the corresponding ISINs of the each bond,
if it exists.
}
\examples{
\dontrun{
  ol_ticker_to_ISIN(openlink_tickers = "TII_0.5_151025",
                     infl_database_path = rminfldbpath)
                   
  ol_ticker_to_ISIN(openlink_tickers = c("TII_0.125_150122", "TII_2_150126"))
}

}
