% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{past_inflation}
\alias{past_inflation}
\title{Historic inflation index data}
\format{
A tibble of 2653 rows and 3 columns
\describe{
\item{inflation_index}{character, the bloomberg code for the inflation index}
\item{index_value}{numeric, the inflation index reading on that date}
\item{date}{date, the date of the inflation reading}
}
}
\source{
bloomberg
}
\usage{
past_inflation
}
\description{
A dataset of historic monthly G7 inflation index readings
}
\keyword{datasets}
