% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_func.R
\name{plot_beta_stability}
\alias{plot_beta_stability}
\title{Plot heat map of beta stability measures}
\usage{
plot_beta_stability(beta_stability, measure = "mean")
}
\arguments{
\item{beta_stability}{a data frame class object in long format, with column names 'window' and 'change_window' and a measure, e.g. 'mean'. It can be generated by the \code{calc_beta_stability_bloomberg} function}

\item{measure}{Object of class character, the measure to be plotted}
}
\value{
rolling beta stability heatmap
}
\description{
For a set of measures of centrality/dispersian of rolling beta series, this function will return a heat map across the window/change window parameters
}
\details{
Plots heatmap of beat stability
}
\examples{
\dontrun{
    plot_beta_stability(beta_stability, measure = "mean")
}


}
