% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beta_func.R
\name{plot_beta_ts}
\alias{plot_beta_ts}
\title{Plot rolling beta series}
\usage{
plot_beta_ts(beta_ts, include_intercept = FALSE)
}
\arguments{
\item{beta_ts}{a data frame class object in long format, with column names 'Date' and 'beta' and optionally 'intercept'. It can be generated by the \code{calc_beta_ts} function}

\item{include_intercept}{Object of class boolean, this will plot the rolling intercept on a second y-axis. This can be helpful for guaging if there is a large level shift.}
}
\value{
rolling beta ts plot
}
\description{
For a given series, this function will return a time series plot of a rolling beta between two variables
}
\details{
Plots rolling beta series for given input
}
\examples{
\dontrun{
    plot_beta_ts(US_oil_beta)
}


}
