% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_supply.R
\name{plot_monthly_supply}
\alias{plot_monthly_supply}
\title{Plot Monthly Linker Supply from Database}
\usage{
plot_monthly_supply(
  infl_link_supply_ts,
  field = "notional",
  start_date = "2018-01-01"
)
}
\arguments{
\item{infl_link_supply_ts}{Object of class data.frame, consisting of inflation linked bond supply statistics}

\item{field}{field to plot, e.g. notional}

\item{start_date}{start date for the series, e.g "2018-01-01"}
}
\value{
plot of inflation-linked bond supply
}
\description{
Returns plotly object of inflation-linked bond supply from inflation database for a given series
}
\details{
Returns a plot of inflation linked bond supply
}
\examples{
\dontrun{
  infl_link_supply_ts <- InflationTools::infl_link_supply_ts_db("US")
  plot_monthly_supply(infl_link_supply_ts)
}

}
