% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linker_supply.R
\name{plot_supply_maturity}
\alias{plot_supply_maturity}
\title{Plot linker Supply by Maturity from Database}
\usage{
plot_supply_maturity(infl_link_supply_ts, start_date = "2017-01-01")
}
\arguments{
\item{infl_link_supply_ts}{Object of class data.frame, consisting of inflation linked bond supply statistics}

\item{start_date}{start date for the series, e.g "2018-01-01"}
}
\value{
plot of inflation-linked bond supply by maturity
}
\description{
Returns plotly object of inflation-linked bond supply from inflation database for a given series
}
\details{
returns plot of inflation linked bond supply
}
\examples{
\dontrun{
  infl_link_supply_ts <- InflationTools::infl_link_supply_ts_db("US")
  plot_supply_maturity(infl_link_supply_ts)
}

}
