% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{real_accrued}
\alias{real_accrued}
\title{Accrued interest calculation (real / inflation unadjusted)}
\usage{
real_accrued(cashflow_table)
}
\arguments{
\item{cashflow_table}{data.table, table of cashflows and discount periods, can be generated by \code{cashflow_table} or \code{linker_cashflow_table}}
}
\value{
numeric, interest accrued on the instrument
}
\description{
Calculate the interest accrued by an instrument since the last cashflow. This is the
real (inflation unadjusted) accrued.
}
\examples{
cash_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
real_accrued(cash_tbl)

}
