% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{real_price}
\alias{real_price}
\title{Real (or clean) price calculation}
\usage{
real_price(cashflow_table, real_yield, payment_frequency, real_accrued)
}
\arguments{
\item{cashflow_table}{data.table, table of cashflows and discount periods, can be generated by \code{cashflow_table}}

\item{real_yield}{numeric, real yield of instrument at reference date}

\item{payment_frequency}{numeric, payment frequency per annum of the instrument eg 2 is semi-annual}

\item{real_accrued}{numeric,interest accrued by an instrument since the last cashflow, can be generated by \code{real_accrued}}
}
\value{
numeric, the clean (or real) price of the instrument
}
\description{
Calculate the real (clean) price of a bond based on a the cashflows and real yield.
}
\examples{
cash_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
real_price(cash_tbl, 
            real_yield = -1,
            payment_frequency = 1,
            real_accrued = 1000)

}
