% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{real_yield}
\alias{real_yield}
\title{Calculate real yield}
\usage{
real_yield(cashflow_table, real_price, payment_frequency, real_accrued)
}
\arguments{
\item{cashflow_table}{data.table, table of cashflows and discount periods, can be generated by \link[InflationTools]{cashflow_table}.}

\item{real_price}{numeric, the real (or clean price) of an inflation linked bond}

\item{payment_frequency}{numeric, payment frequency per annum of the instrument eg 2 is semi-annual}

\item{real_accrued}{numeric,interest accrued by an instrument since the last cashflow, can be generated by \link[InflationTools]{real_accrued}.}
}
\value{
numeric, a yield
}
\description{
Calculate the yield (or real yield) of an instrument. Requires a cashflow table.
}
\examples{
temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
real_yield(temp_cashflow_tbl,
                         145,
                         2,
                         2500)                
                         


}
