% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{rminfldbpath}
\alias{rminfldbpath}
\title{Path of RM inflation database}
\format{
character string, a file path
}
\source{
SH
}
\usage{
rminfldbpath
}
\description{
The specific filepath of the current Reserve Managers inflation database
}
\keyword{datasets}
