% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{rmlmdbpath}
\alias{rmlmdbpath}
\title{Path of RM list management database}
\format{
character string, a file path
}
\source{
SH
}
\usage{
rmlmdbpath
}
\description{
The specific filepath of the current Reserve Managers list management database
}
\keyword{datasets}
