% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_check.R
\name{seasonal_x13_package_install_check}
\alias{seasonal_x13_package_install_check}
\title{Check seasonal and x13binary packages installation}
\usage{
seasonal_x13_package_install_check(install = FALSE)
}
\arguments{
\item{install}{logical, whether the packages should be installed if they are missing}
}
\description{
checks that the x13binary and seasonal packages are installed in the C drive location
on the users computer. Restrictions with Bank software mean that executable files must be
stored in the C drive. Packages can be installed if missing.
}
\examples{
\dontrun{
    seasonal_x13_package_install_check(FALSE)
}
}
