% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_adjustment_cpi.R
\name{seasonality_adjust_projection}
\alias{seasonality_adjust_projection}
\title{Seasonally adjust projected inflation series}
\usage{
seasonality_adjust_projection(
  projected_curve,
  infl_data,
  proj_is_SA = FALSE,
  start_date = Sys.Date(),
  method = "linear",
  outliers_shift = c("ao2009.Oct", "ao2009.Sep")
)
}
\arguments{
\item{projected_curve}{Projected index, expects columns names "proj_infl", "date", "season_factor_cuml", can be generated from
\code{create_linear_unseasonal_proj_curve} or \code{implied_proj_inflation_curve}.}

\item{infl_data}{Object of class 'inflation data'}

\item{proj_is_SA}{Object of class boolean. Is the projection curve seasonally adjusted TRUE or FALSE}

\item{start_date}{Object of class date, the start date for the series}

\item{method}{Default method is set to linear}

\item{outliers_shift}{Vector of characters, of any outliers (ao) or level shifts expected in the data. The \code{seas}
function from the seasonal package will identify these automatically, but they can also be pre-specified if the user
has prior knowledge of the data.}
}
\value{
seasonalally adjusted series
}
\description{
Seasonally adjust projexted inflation index series. This can be used to calculate impact the pricing of inflation linked instruments.
With the the log transformation function, this functions calls the automatic procedures of X-13ARIMA-SEATS to perform a seasonal adjustment on a given index
}
\details{
seasonally_adjust takes an inflation index as input as a long format data frame, queried from inflation database.
}
\examples{
\dontrun{
 seasonality_adjust_projection(projected_curve = exampleUSInflationCurve, proj_is_SA = FALSE)
}

}
