% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_adjustment_cpi.R
\name{seasonality_adjust_projection_db}
\alias{seasonality_adjust_projection_db}
\title{Seasonally adjust projected inflation series from inflation database}
\usage{
seasonality_adjust_projection_db(
  projected_curve,
  projected_index,
  proj_is_SA = FALSE,
  start_date = Sys.Date(),
  method = "linear",
  outliers_shift = c("ao2009.Oct", "ao2009.Sep"),
 
    database_path = "\\\\\\\\markets\\\\DATA\\\\DATA\\\\Offdata\\\\RM\\\\_Data\\\\Inflation\\\\inflation.db"
)
}
\arguments{
\item{projected_curve}{Projected index, expects columns names "proj_infl", "date", "season_factor_cuml", can be generated from
\link[InflationTools]{create_linear_unseasonal_proj_curve}.}

\item{projected_index}{string, inflation index series name, e.g. CPTFEMU Index}

\item{proj_is_SA}{Object of class boolean. Is the projection curve seasonally adjusted TRUE or FALSE}

\item{start_date}{Object of class date, the start date for the series}

\item{method}{Default method is set to linear}

\item{outliers_shift}{Vector of characters, of any outliers (ao) or level shifts expected in the data. The \link[seasonal]{seas}
function from the seasonal package will identify these automatically, but they can also be pre-specified if the user
has prior knowledge of the data.}

\item{database_path}{Path to inflation database, default ////markets//DATA//DATA//Offdata//RM//_Data//Inflation//inflation.db}
}
\value{
seasonalally adjusted series
}
\description{
Seasonally adjust projected inflation index series from inflation database. This can be used to calculate impact the pricing of inflation linked instruments.
With the the log transformation function, this functions calls the automatic procedures of X-13ARIMA-SEATS to perform a seasonal adjustment on a given index
}
\details{
seasonally_adjust takes an inflation index as input as a long format data frame, queried from inflation database.
}
\examples{
\dontrun{
 seasonality_adjust_projection_db(projected_curve = exampleUSInflationCurve, 
                                  projected_index = "CPURNSA Index", 
                                  proj_is_SA = FALSE)
}

}
