% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_adjustment_cpi.R
\name{seasonality_adjustment}
\alias{seasonality_adjustment}
\title{Seasonally adjust inflation index series}
\usage{
seasonality_adjustment(
  infl_data,
  method = "linear",
  outliers_shift = c("ao2009.Oct", "ao2009.Sep")
)
}
\arguments{
\item{infl_data}{Object of class 'inflation data'}

\item{method}{Defaul method is set to linear}

\item{outliers_shift}{Vector of characters, of any outliers (ao) or level shifts expected in the data. The \code{seas}
function from the seasonal package will identify these automatically, but they can also be pre-specified if the user
has prior knowledge of the data.}
}
\value{
seasonalally adjusted series
}
\description{
Seasonally adjusted inflation index series.
With the the log transformation function, this functions calls the automatic procedures of X-13ARIMA-SEATS to perform a seasonal adjustment on a given index
}
\details{
seasonally_adjust takes an inflation index as input as a long format data frame.
Column names should be 'index_value' for the inflation series, and 'date' for the corresponding date
}
\examples{
\dontrun{
  index_to_date <- past_inflation \%>\%
      dplyr::select(date, index_value) \%>\%
      dplyr::mutate(date = lubridate::as_date(date)) 
       
  seasonality_adjustment(infl_data = index_to_date, method = "linear")
}

}
