% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seasonality_adjustment_cpi.R
\name{seasonally_adjust_db}
\alias{seasonally_adjust_db}
\title{Seasonally adjust inflation index series from inflation database}
\usage{
seasonally_adjust_db(
  infl_index,
  end_date = Sys.Date(),
  outliers_shift = c("ao2009.Oct", "ao2009.Sep"),
 
    database_path = "\\\\\\\\markets\\\\DATA\\\\DATA\\\\Offdata\\\\RM\\\\_Data\\\\Inflation\\\\inflation.db"
)
}
\arguments{
\item{infl_index}{Inflation index series name, e.g. CPTFEMU Index}

\item{end_date}{Object of class date, the end date for the series}

\item{outliers_shift}{Vector of characters, of any outliers (ao) or level shifts expected in the data. The \code{seas}
function from the seasonal package will identify these automatically, but they can also be pre-specified if the user
has prior knowledge of the data.}

\item{database_path}{Path to inflation database, default ////markets//DATA//DATA//Offdata//RM//_Data//Inflation//inflation.db}
}
\value{
seasonalally adjusted series
}
\description{
Seasonally adjusted inflation index series from inflation database.
With the the log transformation function, this functions calls the automatic procedures of X-13ARIMA-SEATS to perform a seasonal adjustment on a given index
}
\details{
seasonally_adjust takes an inflation index as input as a long format data frame, queried from inflation database.
}
\examples{

seasonally_adjust_db("CPTFEMU Index")

}
\seealso{
seasonally_adjust
}
