% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cashflow_calculations.R
\name{seasonally_adjusted_real_yield}
\alias{seasonally_adjusted_real_yield}
\title{Calculate seasonally adjusted real yield}
\usage{
seasonally_adjusted_real_yield(
  linker_cashflow_table,
  real_price,
  payment_frequency,
  real_accrued
)
}
\arguments{
\item{linker_cashflow_table}{data.table, table of inflation linked (adjusted) cashflows and discount periods, can be generated by
\link[InflationTools]{linker_cashflow_table}.}

\item{real_price}{numeric, the real (or clean price) of an inflation linked bond}

\item{payment_frequency}{numeric, payment frequency per annum of the instrument eg 2 is semi-annual}

\item{real_accrued}{numeric,interest accrued by an instrument since the last cashflow, can be generated by \link[InflationTools]{real_accrued}.}
}
\value{
numeric, a real yield
}
\description{
Calculate the seasonally adjusted real yield of an inflation linked instrument. Requires a cashflow table, a projected inflation
curve and the seasonally adjusted factors (multiplicative). These can be generated using a combination of functions such as
\link[InflationTools]{linker_cashflow_table} and \link[InflationTools]{implied_proj_inflation_curve_db}.
}
\examples{
temp_cashflow_tbl <- cashflow_table(as.Date("2023-02-21"), 0.1, 2)
linker_cash_tbl <- linker_cashflow_table(temp_cashflow_tbl,
                                         exampleUSInflationCurve,
                                         260)
seasonally_adjusted_real_yield(linker_cash_tbl,
                               145,
                               2,
                               2500)                
                         


}
