% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_data_headers.R
\docType{data}
\name{securities_inflation_model}
\alias{securities_inflation_model}
\title{Path of RM list management database}
\format{
data frame
}
\source{
JA
}
\usage{
securities_inflation_model
}
\description{
The securities config file for the inflation/liquidity risk premia modelling
}
\keyword{datasets}
