% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_inflation_rates.R
\name{swap_implied_inflation_data}
\alias{swap_implied_inflation_data}
\title{Swap Implied Inflation}
\usage{
swap_implied_inflation_data(inflation_index, lag_adjustment = TRUE)
}
\arguments{
\item{inflation_index}{Object of class character, specifying the inflation index}

\item{lag_adjustment}{Logical, whether to adjust the projected curve for the lag in inflation swap market}
}
\value{
data frame swap implied inflation
}
\description{
Returns inflation implied by zero-coupon inflation swaps
}
\details{
In a zero coupon swap, one party pays fixed rate on a notional amount and the other pays floating linked to an inflation index.
This function gives the expected inflation over a number of maturities in percentage terms. The index lag can be left in or removed.
}
\examples{
\dontrun{
  swap_implied_inflation_data("CPURNSA Index")
}

}
