% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_interest_rates.R
\name{swap_implied_interest_rate_curve}
\alias{swap_implied_interest_rate_curve}
\title{Swap implied interest rate curve}
\usage{
swap_implied_interest_rate_curve(interest_rate_curve, ref_date = Sys.Date())
}
\arguments{
\item{interest_rate_curve}{Object of class character, specifying the name of the swap curve index}

\item{ref_date}{Date, the reference date of the curve}
}
\value{
data frame of swap curve with one column per curve
}
\description{
Returns interest rate swap curve from bloomberg
}
\details{
In a zero coupon swap, one party pays fixed rate on a notional amount and the other pays floating linked to an inflation index.
This function gives the expected inflation over a number of maturities in percentage terms. The index lag can be left in or removed.
}
\examples{
\dontrun{
  swap_implied_interest_rate_curve("SOFR")
}

}
