% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/implied_interest_rates.R
\name{swap_implied_interest_rate_curve_historic}
\alias{swap_implied_interest_rate_curve_historic}
\title{Swap implied interest rate curve for multiple dates}
\usage{
swap_implied_interest_rate_curve_historic(
  interest_rate_curve,
  from_date = Sys.Date() - 20,
  to_date = Sys.Date()
)
}
\arguments{
\item{interest_rate_curve}{Object of class character, specifying the name of the swap curve index}

\item{from_date}{Date, the earliest date to get the curve}

\item{to_date}{Date, the final date to get the curve}
}
\value{
data frame of swap curve
}
\description{
Returns interest rate swap curve from bloomberg
}
\details{
In a zero coupon swap, one party pays fixed rate on a notional amount and the other pays floating linked to an inflation index.
This function gives the expected inflation over a number of maturities in percentage terms. The index lag can be left in or removed.
}
\examples{
\dontrun{
  swap_implied_interest_rate_curve_historic("SOFR")
}

}
