% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmdashboard_functions.R
\name{ZC_risk_dashboard}
\alias{ZC_risk_dashboard}
\title{Openlink report zero-coupon risk calculation}
\usage{
ZC_risk_dashboard(
  instrument_ticker,
  maturity_date,
  nominal,
  currency,
  infl_database_path = rminfldbpath
)
}
\arguments{
\item{instrument_ticker}{character, the openlink ticker, normally a free text format for swaps. It is possible to include
a base or reference CPI in the ticker eg \code{"New zc swap [258.27]"} to improve calculation.}

\item{maturity_date}{date, the date of the maturity ie final cashflow}

\item{nominal}{numeric, the face value or nominal of the instrument traded}

\item{currency}{character, the currency denomination of the instrument}

\item{infl_database_path}{Path class character to inflation database}
}
\value{
list, includes DV01, IE01 and any error messages
}
\description{
Calculate the risk associated with zero-coupon inflation linked swap (ILS) holdings from information
found in the openlink report
}
\details{
Runs using the more limited information provided in the daily openlink P&L reports. Is supplemented
by inflation and list management databases for static data, so it doesn't need to connect to bloomberg.
Probably shouldn't be used outside of openlink with \link[InflationTools]{zero_coupon_inflation_swap} doing the
same calculations with less input required.
}
\examples{
ZC_risk_dashboard(instrument_ticker = "example ZC swap [257.4]",
                   maturity_date = as.Date("2031-01-01"),
                   nominal = 10000000, 
                   currency = "EUR")


}
