% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_coupon_swap.R
\name{zero_coupon_inflation_swap}
\alias{zero_coupon_inflation_swap}
\title{Zero coupon inflation swap}
\usage{
zero_coupon_inflation_swap(
  ref_date,
  maturity_date,
  fixed_rate,
  notional,
  projected_inflation_curve,
  inflation_lag = 3,
  reference_cpi = 259.918,
  settle_convention = 2,
  start_date = NULL,
  interpolation = "daily",
  day_count = "ACT/ACT"
)
}
\arguments{
\item{ref_date}{date, the trade date of the}

\item{maturity_date}{date, the maturity date for the series}

\item{fixed_rate}{Class numeric, the fixed payment of the ZC inflation swap}

\item{notional}{Class numeric, the notional amount of the ZC inflation swap}

\item{projected_inflation_curve}{data.table, curve of projected inflation with columns \code{date} and \code{proj_infl},
eg can be generated by \link[InflationTools]{implied_proj_inflation_curve}}

\item{inflation_lag}{Class numeric, the inflation lag for the index, default is 3 (months)}

\item{reference_cpi}{Class numeric, reference CPI value}

\item{settle_convention}{numeric, the days until the instrument will be settled}

\item{start_date}{class,}

\item{interpolation}{Class character, can be daily or monthly, inflation interpolation method}

\item{day_count}{Class character, defaults to ACT/ACT, day count convention}
}
\value{
list of swap characteristics
}
\description{
Returns zero coupon inflation swap information, including NPV, fixed/floating payments, DV01 and IE01
}
\details{
Zero coupon inflation swap calculations, for a given start, maturity, fixed rate, notional, CPI reference and projected inflation curve
}
\examples{

zero_coupon_inflation_swap(ref_date = Sys.Date(), maturity_date = "2025-04-02",
                           fixed_rate = 2.71, notional = 10000000,
                           projected_inflation_curve = exampleUSInflationCurve)


}
